refactor(agent-core): 重组 smoke 分层入口
