+++ packages/agent-core/src/smoke/runtime-parity-smoke.ts
@@ -1,7 +1,1 @@
-import { runRuntimeParitySmoke } from './cases/runtime/runtime-parity-case.js';
-
-runRuntimeParitySmoke().catch((error: unknown) => {
-  const message = error instanceof Error ? error.message : String(error);
-  console.error(`runtime parity smoke failed: ${message}`);
-  process.exitCode = 1;
-});
+import './runtime/runtime-parity-smoke.js';
