+++ packages/agent-core/src/smoke/runtime/runtime-parity-smoke.ts
@@ -1,0 +1,7 @@
+import { runRuntimeParitySmoke } from '../cases/runtime/runtime-parity-case.js';
+
+runRuntimeParitySmoke().catch((error: unknown) => {
+  const message = error instanceof Error ? error.message : String(error);
+  console.error(`runtime parity smoke failed: ${message}`);
+  process.exitCode = 1;
+});
