+++ packages/agent-core/src/smoke/shared/index.ts
@@ -1,0 +1,4 @@
+export * from './print.js';
+export * from './env.js';
+export * from './demo-tool.js';
+export * from './runtime.js';
