+++ packages/agent-core/src/smoke/shared/print.ts
@@ -1,0 +1,4 @@
+export function printSmokeSection(title: string, payload: unknown): void {
+  console.log(`=== ${title} ===`);
+  console.log(JSON.stringify(payload, null, 2));
+}
